trapped cards algorithm adjust
take the depths of the set of shallowest trapped cards in each column.  assume that every card above must be moved aside in order to un-trap that card.  Right now, it takes 1 move (minimum) per card to set it aside.
So - taking that, we can subtract those moves from the maximum depth, combined with the remaining cards on the deck in order to determine if we should go further down this tree - it means we terminate trees earlier in the process, hopefully speeding calculations.
Tableau gets a new trappedDepths and trappedCardHeight methods for these calculations.
TableauMoveIterator now uses those in depthFunction to terminate recursion.
StagedDepthFirstSolver also now uses them in flushDeepTrees when the max depth has been reset.