critical fix
I was not properly handing the separation between descendFor and queueLeaves - some (many?) of the queued trees would be getting removed because they had no children.
Solution (temporarily) is to keep the move iteration state in the MoveTree rather than MoveState.  Need to move it back there, but...
Now, a tree is only retired if it both has no moves left and no children tree nodes.